+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### master
 
+- add `fetchOne` to fetch 0 or 1 rows, throwing for results with >1 rows
+- rename `fetchOne` to `fetchExactlyOne`
 - add `orderBy` to sort results
 - add `selectAs` to rename selected columns
 - rename `selectAs` to `selectAsJson`